Create an abstraction around SPM mutation to remove it out of BSVM.
