+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/viewmodels/BaseSheetViewModel.kt
@@ -14,17 +14,15 @@ import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
 import com.stripe.android.model.CardBrand
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCode
-import com.stripe.android.model.PaymentMethodUpdateParams
 import com.stripe.android.model.SetupIntent
 import com.stripe.android.payments.paymentlauncher.PaymentResult
 import com.stripe.android.paymentsheet.LinkHandler
 import com.stripe.android.paymentsheet.NewOrExternalPaymentSelection
-import com.stripe.android.paymentsheet.PaymentOptionsItem
-import com.stripe.android.paymentsheet.PaymentOptionsViewModel
 import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.paymentsheet.PaymentSheet.PaymentMethodLayout
 import com.stripe.android.paymentsheet.PaymentSheetViewModel
 import com.stripe.android.paymentsheet.PrefsRepository
+import com.stripe.android.paymentsheet.SavedPaymentMethodMutator
 import com.stripe.android.paymentsheet.analytics.EventReporter
 import com.stripe.android.paymentsheet.analytics.PaymentSheetAnalyticsListener
 import com.stripe.android.paymentsheet.model.MandateText
@@ -33,21 +31,16 @@ import com.stripe.android.paymentsheet.model.PaymentSelection.CustomerRequestedS
 import com.stripe.android.paymentsheet.navigation.NavigationHandler
 import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen
 import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen.AddAnotherPaymentMethod
-import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen.AddFirstPaymentMethod
 import com.stripe.android.paymentsheet.repositories.CustomerRepository
-import com.stripe.android.paymentsheet.state.CustomerState
 import com.stripe.android.paymentsheet.state.GooglePayState
 import com.stripe.android.paymentsheet.state.WalletsProcessingState
 import com.stripe.android.paymentsheet.state.WalletsState
 import com.stripe.android.paymentsheet.ui.DefaultAddPaymentMethodInteractor
-import com.stripe.android.paymentsheet.ui.EditPaymentMethodViewInteractor
 import com.stripe.android.paymentsheet.ui.HeaderTextFactory
 import com.stripe.android.paymentsheet.ui.ModifiableEditPaymentMethodViewInteractor
-import com.stripe.android.paymentsheet.ui.PaymentMethodRemovalDelayMillis
 import com.stripe.android.paymentsheet.ui.PaymentSheetTopBarState
 import com.stripe.android.paymentsheet.ui.PaymentSheetTopBarStateFactory
 import com.stripe.android.paymentsheet.ui.PrimaryButton
-import com.stripe.android.ui.core.cbc.CardBrandChoiceEligibility
 import com.stripe.android.ui.core.elements.CvcConfig
 import com.stripe.android.ui.core.elements.CvcController
 import com.stripe.android.uicore.utils.combine
@@ -55,7 +48,6 @@ import com.stripe.android.uicore.utils.combineAsStateFlow
 import com.stripe.android.uicore.utils.mapAsStateFlow
 import com.stripe.android.uicore.utils.stateFlowOf
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
@@ -72,17 +64,17 @@ import kotlin.coroutines.CoroutineContext
 @Suppress("TooManyFunctions")
 internal abstract class BaseSheetViewModel(
     application: Application,
-    internal val config: PaymentSheet.Configuration,
-    internal val eventReporter: EventReporter,
-    protected val customerRepository: CustomerRepository,
+    val config: PaymentSheet.Configuration,
+    val eventReporter: EventReporter,
+    val customerRepository: CustomerRepository,
     protected val prefsRepository: PrefsRepository,
-    protected val workContext: CoroutineContext = Dispatchers.IO,
+    val workContext: CoroutineContext = Dispatchers.IO,
     protected val logger: Logger,
     val savedStateHandle: SavedStateHandle,
     val linkHandler: LinkHandler,
     val linkConfigurationCoordinator: LinkConfigurationCoordinator,
     private val headerTextFactory: HeaderTextFactory,
-    private val editInteractorFactory: ModifiableEditPaymentMethodViewInteractor.Factory
+    val editInteractorFactory: ModifiableEditPaymentMethodViewInteractor.Factory
 ) : AndroidViewModel(application) {
 
     protected var mostRecentError: Throwable? = null
@@ -95,22 +87,6 @@ internal abstract class BaseSheetViewModel(
 
     private val _supportedPaymentMethodsFlow = MutableStateFlow<List<PaymentMethodCode>>(emptyList())
     val supportedPaymentMethodsFlow: StateFlow<List<PaymentMethodCode>> = _supportedPaymentMethodsFlow
-
-    protected var customer: CustomerState?
-        get() = savedStateHandle[SAVED_CUSTOMER]
-        set(value) {
-            savedStateHandle[SAVED_CUSTOMER] = value
-        }
-
-    /**
-     * The list of saved payment methods for the current customer.
-     * Value is null until it's loaded, and non-null (could be empty) after that.
-     */
-    internal val paymentMethods: StateFlow<List<PaymentMethod>> = savedStateHandle
-        .getStateFlow<CustomerState?>(SAVED_CUSTOMER, null)
-        .mapAsStateFlow { state ->
-            state?.paymentMethods ?: emptyList()
-        }
 
     val navigationHandler: NavigationHandler = NavigationHandler { poppedScreen ->
         analyticsListener.reportPaymentSheetHidden(poppedScreen)
@@ -193,38 +169,7 @@ internal abstract class BaseSheetViewModel(
         !isProcessing && !isEditing
     }
 
-    private val paymentOptionsItemsMapper: PaymentOptionsItemsMapper by lazy {
-        PaymentOptionsItemsMapper(
-            paymentMethods = paymentMethods,
-            googlePayState = googlePayState,
-            isLinkEnabled = linkHandler.isLinkEnabled,
-            isNotPaymentFlow = this is PaymentOptionsViewModel,
-            nameProvider = ::providePaymentMethodName,
-            isCbcEligible = { paymentMethodMetadata.value?.cbcEligibility is CardBrandChoiceEligibility.Eligible }
-        )
-    }
-
-    internal fun providePaymentMethodName(code: PaymentMethodCode?): String {
-        return code?.let {
-            paymentMethodMetadata.value?.supportedPaymentMethodForCode(code)
-        }?.displayName?.resolve(getApplication()).orEmpty()
-    }
-
-    val paymentOptionsItems: StateFlow<List<PaymentOptionsItem>> = paymentOptionsItemsMapper()
-
-    private val canEdit: StateFlow<Boolean> = paymentOptionsItems.mapAsStateFlow { items ->
-        val paymentMethods = items.filterIsInstance<PaymentOptionsItem.SavedPaymentMethod>()
-        if (config.allowsRemovalOfLastSavedPaymentMethod) {
-            paymentMethods.isNotEmpty()
-        } else {
-            if (paymentMethods.size == 1) {
-                // We will allow them to change card brand, but not delete.
-                paymentMethods.first().isModifiable
-            } else {
-                paymentMethods.size > 1
-            }
-        }
-    }
+    val savedPaymentMethodMutator: SavedPaymentMethodMutator = SavedPaymentMethodMutator.create(this)
 
     val topBarState: StateFlow<PaymentSheetTopBarState> = combineAsStateFlow(
         navigationHandler.currentScreen.mapAsStateFlow { it.sheetScreen },
@@ -232,7 +177,7 @@ internal abstract class BaseSheetViewModel(
         paymentMethodMetadata.mapAsStateFlow { it?.stripeIntent?.isLiveMode ?: true },
         processing,
         editing,
-        canEdit,
+        savedPaymentMethodMutator.canEdit,
         PaymentSheetTopBarStateFactory::create,
     )
 
@@ -247,7 +192,7 @@ internal abstract class BaseSheetViewModel(
 
     init {
         viewModelScope.launch {
-            canEdit.collect { canEdit ->
+            savedPaymentMethodMutator.canEdit.collect { canEdit ->
                 if (!canEdit && editing.value) {
                     toggleEditing()
                 }
@@ -255,7 +200,7 @@ internal abstract class BaseSheetViewModel(
         }
 
         viewModelScope.launch {
-            paymentMethods.collect { paymentMethods ->
+            savedPaymentMethodMutator.paymentMethods.collect { paymentMethods ->
                 if (paymentMethods.isEmpty() && editing.value) {
                     toggleEditing()
                 }
@@ -301,6 +246,12 @@ internal abstract class BaseSheetViewModel(
                 _mandateText.value = null
             }
         }
+    }
+
+    internal fun providePaymentMethodName(code: PaymentMethodCode?): String {
+        return code?.let {
+            paymentMethodMetadata.value?.supportedPaymentMethodForCode(code)
+        }?.displayName?.resolve(getApplication()).orEmpty()
     }
 
     protected fun transitionToFirstScreen() {
@@ -431,15 +382,6 @@ internal abstract class BaseSheetViewModel(
         _contentVisible.value = visible
     }
 
-    fun removePaymentMethod(paymentMethod: PaymentMethod) {
-        val paymentMethodId = paymentMethod.id ?: return
-
-        viewModelScope.launch(workContext) {
-            removeDeletedPaymentMethodFromState(paymentMethodId)
-            removePaymentMethodInternal(paymentMethodId)
-        }
-    }
-
     private fun updateCvcFlows(selection: PaymentSelection?) {
         if (selection is PaymentSelection.Saved && selection.paymentMethod.type == PaymentMethod.Type.Card) {
             _cvcControllerFlow.value = CvcController(
@@ -451,166 +393,6 @@ internal abstract class BaseSheetViewModel(
                     _cvcRecollectionCompleteFlow.value = it
                 }
             }
-        }
-    }
-
-    private suspend fun removePaymentMethodInternal(paymentMethodId: String): Result<PaymentMethod> {
-        // TODO(samer-stripe): Send 'unexpected_error' here
-        val currentCustomer = customer ?: return Result.failure(
-            IllegalStateException(
-                "Could not remove payment method because CustomerConfiguration was not found! Make sure it is " +
-                    "provided as part of PaymentSheet.Configuration"
-            )
-        )
-
-        val currentSelection = (selection.value as? PaymentSelection.Saved)?.paymentMethod?.id
-        val didRemoveSelectedItem = currentSelection == paymentMethodId
-
-        if (didRemoveSelectedItem) {
-            // Remove the current selection. The new selection will be set when we're computing
-            // the next PaymentOptionsState.
-            updateSelection(null)
-        }
-
-        return customerRepository.detachPaymentMethod(
-            CustomerRepository.CustomerInfo(
-                id = currentCustomer.id,
-                ephemeralKeySecret = currentCustomer.ephemeralKeySecret
-            ),
-            paymentMethodId
-        )
-    }
-
-    private fun removeDeletedPaymentMethodFromState(paymentMethodId: String) {
-        val currentCustomer = customer ?: return
-
-        customer = currentCustomer.copy(
-            paymentMethods = currentCustomer.paymentMethods.filter {
-                it.id != paymentMethodId
-            }
-        )
-
-        if (mostRecentlySelectedSavedPaymentMethod.value?.id == paymentMethodId) {
-            savedStateHandle[SAVED_PM_SELECTION] = null
-        }
-
-        if ((selection.value as? PaymentSelection.Saved)?.paymentMethod?.id == paymentMethodId) {
-            savedStateHandle[SAVE_SELECTION] = null
-        }
-
-        val shouldResetToAddPaymentMethodForm = paymentMethods.value.isEmpty() &&
-            navigationHandler.currentScreen.value is PaymentSheetScreen.SelectSavedPaymentMethods
-
-        if (shouldResetToAddPaymentMethodForm) {
-            navigationHandler.resetTo(
-                listOf(AddFirstPaymentMethod(interactor = DefaultAddPaymentMethodInteractor.create(this)))
-            )
-        }
-    }
-
-    fun modifyPaymentMethod(paymentMethod: PaymentMethod) {
-        val canRemove = if (config.allowsRemovalOfLastSavedPaymentMethod) {
-            true
-        } else {
-            paymentOptionsItems.value.filterIsInstance<PaymentOptionsItem.SavedPaymentMethod>().size > 1
-        }
-
-        navigationHandler.transitionTo(
-            PaymentSheetScreen.EditPaymentMethod(
-                editInteractorFactory.create(
-                    initialPaymentMethod = paymentMethod,
-                    eventHandler = { event ->
-                        when (event) {
-                            is EditPaymentMethodViewInteractor.Event.ShowBrands -> {
-                                eventReporter.onShowPaymentOptionBrands(
-                                    source = EventReporter.CardBrandChoiceEventSource.Edit,
-                                    selectedBrand = event.brand
-                                )
-                            }
-                            is EditPaymentMethodViewInteractor.Event.HideBrands -> {
-                                eventReporter.onHidePaymentOptionBrands(
-                                    source = EventReporter.CardBrandChoiceEventSource.Edit,
-                                    selectedBrand = event.brand
-                                )
-                            }
-                        }
-                    },
-                    displayName = providePaymentMethodName(paymentMethod.type?.code),
-                    removeExecutor = { method ->
-                        removePaymentMethodInEditScreen(method)
-                    },
-                    updateExecutor = { method, brand ->
-                        modifyCardPaymentMethod(method, brand)
-                    },
-                    canRemove = canRemove,
-                )
-            )
-        )
-    }
-
-    private suspend fun removePaymentMethodInEditScreen(paymentMethod: PaymentMethod): Throwable? {
-        val paymentMethodId = paymentMethod.id!!
-        val result = removePaymentMethodInternal(paymentMethodId)
-
-        if (result.isSuccess) {
-            viewModelScope.launch(workContext) {
-                navigationHandler.pop()
-                delay(PaymentMethodRemovalDelayMillis)
-                removeDeletedPaymentMethodFromState(paymentMethodId = paymentMethodId)
-            }
-        }
-
-        return result.exceptionOrNull()
-    }
-
-    private suspend fun modifyCardPaymentMethod(
-        paymentMethod: PaymentMethod,
-        brand: CardBrand
-    ): Result<PaymentMethod> {
-        // TODO(samer-stripe): Send 'unexpected_error' here
-        val currentCustomer = customer ?: return Result.failure(
-            IllegalStateException(
-                "Could not update payment method because CustomerConfiguration was not found! Make sure it is " +
-                    "provided as part of PaymentSheet.Configuration"
-            )
-        )
-
-        return customerRepository.updatePaymentMethod(
-            customerInfo = CustomerRepository.CustomerInfo(
-                id = currentCustomer.id,
-                ephemeralKeySecret = currentCustomer.ephemeralKeySecret
-            ),
-            paymentMethodId = paymentMethod.id!!,
-            params = PaymentMethodUpdateParams.createCard(
-                networks = PaymentMethodUpdateParams.Card.Networks(
-                    preferred = brand.code
-                ),
-                productUsageTokens = setOf("PaymentSheet"),
-            )
-        ).onSuccess { updatedMethod ->
-            customer = currentCustomer.copy(
-                paymentMethods = currentCustomer.paymentMethods.map { savedMethod ->
-                    val savedId = savedMethod.id
-                    val updatedId = updatedMethod.id
-
-                    if (updatedId != null && savedId != null && updatedId == savedId) {
-                        updatedMethod
-                    } else {
-                        savedMethod
-                    }
-                }
-            )
-
-            handleBackPressed()
-
-            eventReporter.onUpdatePaymentMethodSucceeded(
-                selectedBrand = brand
-            )
-        }.onFailure { error ->
-            eventReporter.onUpdatePaymentMethodFailed(
-                selectedBrand = brand,
-                error = error,
-            )
         }
     }
 
@@ -666,7 +448,6 @@ internal abstract class BaseSheetViewModel(
     abstract fun onError(error: String? = null)
 
     companion object {
-        internal const val SAVED_CUSTOMER = "customer_info"
         internal const val SAVE_SELECTION = "selection"
         internal const val SAVED_PM_SELECTION = "saved_selection"
         internal const val SAVE_PROCESSING = "processing"
