+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/SavedPaymentMethodMutatorTest.kt
@@ -1,0 +1,9 @@
+package com.stripe.android.paymentsheet
+
+import org.junit.Test
+
+class SavedPaymentMethodMutatorTest {
+    @Test
+    fun foo() {
+    }
+}
